fix: 등록할 때 count도 올라가도록 수정
